Add a commit template
Explanation:
In order to use a common style when creating commits a commit template will be
helpful espcially to new developers.

The commit template can be activated with the following command:
git config commit.template .git-commit-template

When this is done any new commit will use the template as the default commit
message.