sdap_save_user: save original primary GID of subdomain users
If ID mapping is enabled we use magic private groups (MPG) for
subdomains, i.e. the UID and the primary GID of the user will have the
same numerical value. As a consequence the information about the
original primary group might get lost because neither in AD domains nor
on a typical UNIX system the user is an explicit member of it's primary
group.

With this patch the mapped GID or the original primary group is saved in
the cached user object under a new attribute.

Fixes https://fedorahosted.org/sssd/ticket/2027