AD: Download master domain info when enumerating
https://fedorahosted.org/sssd/ticket/2068

With the current design, downloading master domain data was tied to
subdomains refresh, triggered by responders. But because enumeration is
a background task that can't be triggered on its own, we can't rely on
responders to download the master domain data and we need to check the
master domain on each enumeration request.