krb5: Ingnore unknown expansion sequences
Recently support was added to use also libkrb5 style expansions that
uses a %{varname} type of template.

There are a number of templates we do not care/can't expand in sssd.
The current code misses tests and failed to properly preserve some of
the templates we do not want to handle.

Addiotionally in order to be future proof this patch treats unknown
templates as pass-through templates and defer any error checking to
libkrb5, so that sssd is consistent with how kinit would behave.

Resolves:
https://fedorahosted.org/sssd/ticket/2076