mmap_cache: Do not remove record from chain twice
It is not very likely, that record will have the same hash1 and hash2, but it
is possible. In this situation, it does not make sense to remove record twice.

Function sss_mc_rm_rec_from_chain was not robust and sssd_nss could crash
in this situation. It was only possible if record was alone in chain.

Resolves:
https://fedorahosted.org/sssd/ticket/2049