IPA: Ignore dns_discovery_domain in server mode
https://fedorahosted.org/sssd/ticket/2079

If the dns_discovery_domain is set in the server mode, then the current
failover code will use it to discover the AD servers as well. This patch
resets the discovery domain unless the admin configured SRV resolution
for IPA servers manually. In the case he did, we try to warn him that
service discovery of AD servers will most likely fail.