sdap_idmap: properly handle ranges for external mappings
Currently we relied on the fact that external ID mapping is used as
default fallback in case of an error and did not properly add subdomains
with external ID mapping to the idmap library. If debugging is enabled
this leads to irritating debug messages for every user or group lookup.

With this patch this subdomains are added to the idmap library.

Fixes https://fedorahosted.org/sssd/ticket/2105