sdap_get_generic_ext_send: check if we a re still connected
At the beginning of a LDAP request we check if we are connecte and have
a valid sdap handle. But for some requests more than one LDAP operation,
typically a search, is needed. Due to the asynchronous handling of LDAP
request it might be possible that a second request might detect a server
error and close the connection while the first request just finished one
LDAP search and wants to start a new LDAP search.

This patch tries to make sure that there is a valid sdap handle before
sending a LDAP search to the server.

Fixes https://fedorahosted.org/sssd/ticket/2126