Add unconditional online callbacks
Currently online callbacks are only executed if the backend was offline
before. This patch add a new class of callback which are always called
if the backend gets a request to go online.

They can be used e.g. to reset timeouts until a more sophisticated method
(OpenLMI, sssctl) is available.