IPA: add trusted domains with missing idrange
If the forest root of a trusted forest is managing POSIX IDs for its
users and groups the same is assumed for all member domains in the
forest which do not have explicitly have an idrange set.

To reflect this SSSD will create the matching ranges automatically.

Fixes https://fedorahosted.org/sssd/ticket/2101