AD: Refresh subdomain data structures on startup
Previously, if no changes were done to the list of subdomains, the SSSD
didn't update its list of sdap_domain mappings for the new subdomain.
This resulted in errors as no id_ctx was present for the subdomain
during lookup.

This patch moves the block of code performed during update to a function
of its own and calls it during provider initialization as well.