NSS: Fix warning access array with index then check
Reported by:cppcheck

Defensive programming: The variable 'i' is used as an array index
before it is checked that is within limits. This can mean that the array might
be accessed out of bounds.

This patch eorder condition such as '(a[i] && i <blen) to
(i < blen && a[i]). That way the array will not be accessed if the index
is out of limits.

Reviewed-by: Michal Žídek <mzidek@redhat.com>