config API: prepend source dir search path for tests
Instead of appending the search patch in the source directory should be
prepended. Otherwise the test might find files installed in the default
paths of the system first.

As a result the compiled python files in the build directory must be
remove in the clean target to make 'make distcheck' pass.

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>