AUTOMAKE: Do not include makefile generated files into tarball
sssd.service was assigned to the dist_systemdunit_DATA variable.
Automake will install this file into the systemd unit directory after building
it if necessary. Automake will also include this generated file in the tarball.
As a result, when building sssd from the tarball, the paths will be not
recreated.

The files in DATA primaries are added as dependencies to the all target
via the internal all-am target. If sssd.service doesn’t exist, make will look
for a rule to build it. Since there is such a rule, make will simply execute
that rule when I build the all target.

Resolves:
https://fedorahosted.org/sssd/ticket/2314

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>