KRB5: Print a verbose error message on failure reading the keytab
krb5_kt_resolve() returns 0 when a non-existent keytab is read, which
means there was no FATAL-level DEBUG message printed to the user in case
the keytab was missing completely and users had to enable more verbose
debugging to diagnose failure to start up.

This patch adds both the verbose DEBUG message as well as a syslog
message.

Reviewed-by: Michal Židek <mzidek@redhat.com>