SBUS: Allow registering paths with fallback
Some interfaces we support, such as domain might not be bound to a
single path, but rather anything under a path 'directory'. This patch
allows the SBUS to register a fallback path that would route any
messages under a given anchor to a handler of an interface.

The fallback interface is denoted with a trailing wildcard:
    /org/sssd/anchor/*

Reviewed-by: Pavel Březina <pbrezina@redhat.com>