resolv: Do not try to free addrinfo in case of error
We should call freeaddrinfo only in situation if getaddrinfo succeeds.
It is not specified in RFC2553 the behaviour of calling freeaddrinfo
with the NULL argument. It is more portable to do not rely on undocumented
behaviour of glibc.

http://www.ietf.org/rfc/rfc2553.txt [Page 28]

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>
(cherry picked from commit 7aaf9138ccf7ab73883f28cfafd6e508d29ffa67)