BUILD: Disable dbus tests when running distcheck
https://fedorahosted.org/sssd/ticket/2291

The dbus tests that mock an sbus server were failing when make distcheck
was ran by a user logged in through the SSSD.

The reason was that the libtool wrapper around the test library alters
the LD_LIBRARY_PATH and as a consequence, the standard getpwuid_r() calls
the dbus server performs would load the in-tree NSS library and not the
system one.  The-in tree library would then attempt to talk to an in-tree
NSS socket, fail, which would fail the getpwuid_r call with an error such as:

"""
    Could not get password database information for UID of current process:
    User "???" unknown or no memory to allocate password entry
"""

This patch adds a new configure-time option called --enable-dbus-tests
that is enabled by default and disabled during distcheck. When the
option is disabled, the tests that require a mocked dbus server are not
compiled at all.