Exit offline mode only if server is available.
This patch adds periodic check to test if
backend can exit offline mode and only
marks backend as not offline if server
for the service is available.

Prior to this patch we marked backend as not
offline if the offline_timeout was reached
without checking for the server availability
and when the next request failed again
we switched back to the offline mode.
This caused significant slowdowns in some
edge cases.

Fixes:
https://fedorahosted.org/sssd/ticket/2355

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>
Reviewed-by: Simo Sorce <simo@redhat.com>